Refinamento de codigo nos controladores

+++ modelos/modelos.go
@@ -51,10 +51,10 @@ type Carrinho struct {
 }
 
 type Items_Carrinho struct {
-	Id          int     `json:"id" gorm:"primaryKey"`
-	Id_Carrinho int     `json:"id_carrinho" gorm:"not null"`
-	ID_Produto  int     `json:"id_produto" gorm:"not null"`
-	Quantidade  float64 `json:"quantidade" gorm:"not null"`
+	Id          int `json:"id" gorm:"primaryKey"`
+	Id_Carrinho int `json:"id_carrinho" gorm:"not null"`
+	ID_Produto  int `json:"id_produto" gorm:"not null"`
+	Quantidade  int `json:"quantidade" gorm:"not null"`
 }
 
 type Pedido struct {
